Minor Features
The game now shows a brief tutorial before starting the timer. This gives the user a chance to understand what's going on before the timer fires them for slacking.

The logs now have a class for ease of use, and there's basic functionality for reading them from a list. Needs improvement, however.